requirements.txt
openai
psycopg2
langchain
python-environ
